- Automatisk urlhåndtering i veggkommentarer - Diverse andre småjusteringer
git-svn-id: https://subversion.kantega.no/svn/aksess/projects/aksess-modules/trunk@29811 497bca9e-fae7-0310-9681-f2327195b1a6

+++ modules/forum/src/java/no/kantega/forum/dao/ForumDao.java
@@ -6,6 +6,7 @@ import org.springframework.orm.hibernate3.HibernateCallback;
 import org.apache.log4j.Logger;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.*;
 import java.sql.SQLException;
 
@@ -165,7 +166,7 @@ public class ForumDao {
 
                 queryThreadIds.setString(0, userId);
 
-                List<BigDecimal> threadIds = queryThreadIds.list();
+                List<BigInteger> threadIds = queryThreadIds.list();
                 if (threadIds.size() == 0) {
                     return new ArrayList<ForumThread>();
                 }
@@ -193,8 +194,8 @@ public class ForumDao {
                 }
 
                 for (int i = 0; i < threadIds.size(); i++) {
-                    BigDecimal tId = threadIds.get(i);
-                    query.setBigDecimal(i, tId);
+                    BigInteger tId = threadIds.get(i);
+                    query.setBigInteger(i, tId);
                 }
 
                 List<ForumThread> threads = query.list();
